+++ classes/hvp_toolbox.php
@@ -25,6 +25,8 @@
  */
 namespace theme_adaptable;
 
+use core\url;
+
 /**
  * Trait theme_adaptable_core_h5p_renderer.
  *
@@ -54,14 +56,14 @@ trait hvp_toolbox {
      *
      * @param string $content Content.
      *
-     * @return moodle_url the URL.
+     * @return url the URL.
      */
     protected function get_style_url($content) {
         global $CFG;
 
         $syscontext = \context_system::instance();
         $itemid = md5($content);
-        return \moodle_url::make_file_url(
+        return url::make_file_url(
             "$CFG->wwwroot/pluginfile.php",
             "/$syscontext->id/theme_adaptable/hvp/$itemid/themehvp.css"
         );