+++ classes/output/core_renderer_maintenance.php
@@ -32,6 +32,6 @@ namespace theme_adaptable\output;
 /**
  * The core maintenance renderer.
  */
-class core_renderer_maintenance extends \core_renderer_maintenance {
+class core_renderer_maintenance extends \core\output\core_renderer_maintenance {
     use core_renderer_toolbox, core_renderer_layout;
 }